Do not show notifications about messages that have already been shown by the approver.
This patch actually splits the handleIncomingMessage() function into
two functions, one that shows the message on the window and one that
shows a notification, and connects them independently on the incoming
message signal. This way, while the chat is initializing, new incoming
messages are shown in the notification area, but not on the chat view.
When the chat view is ready, it processes all the queued messages, which
include the ones that were shown earlier on the approver and the ones that
were shown while the view was initializing.