Re-establish chat after going offline and back online part 1
this patch is in preparation for closing bug #270725.
I basically created some setup methods for connecting the signals and slots for the textchannels, chat tab
and the contact model of chatwidget.
Plus the search for duplicate tabs is no longer done by matching the incomingTextChannel, but the targetId
of the channel.