Simplify some bits of the chat plasmoid's code
Simplify the ConversationDelegate by actually using a Button and a QItemWidget, don't deal with plasma themes directly for no reason (if there was, I didn't see it).
Unify the checked state of the chat views. Now there's not the problem when you press "Esc" that the Chat disappeared but the button remained pressed.

REVIEW: 104543