Implement Searchexpansion filter
This simply takes the whole message and passes it to KUriFilter
to 'filter'. If it was succesful (meaning message was a web
shortcut), it replaces it with the filtered uri.

Note: The stuck in Url doesn't go through UrlFilter again, so
it won't be clickable. Should it matter, since this is outgoing
anyway?