Handle connection states and print a service message when it changes
Until now we print "Connected" and "Connection closed" strings when the channel is connected or invalidated. This patch watches for channel connection and prints the corresponding presence as a service message, ie. "You are now offline" and "You are now marked as %1" instead of those "Connected"/"Connection closed" strings.

Reviewed-by: David Edmundson
BUG: 304092
FIXED-IN: 0.5