Implementt Youtube Filter
Currently, it embeds it using HTML5. This won't work with the chat
plasmoid, but it appears to ignore it gracefully, so I'll let it be.

And besides, I think it's better that way, because giving a preview
of a video that you can't play would probably only strive to anger
users.

Conflicts:
	filters/CMakeLists.txt
	tests/message-processor-basic-tests.cpp
	tests/message-processor-basic-tests.h