Implement Bugzilla Filter
I tried to use the bugzilla API to show some useful information,
but KIO doesn't let me do any of network requests synchronously.

So, for now this is going to be just embed a link to the bugzilla
page. I'll come back and make it nice when I have an asynchronous
plugin system.