Call the onContactPresenceChange directly without connection checks
The connection status at this point is still Tp::ConnectionStatusDisconnected
and so the check for connection being Tp::ConnectionStatusConnected
would never pass, which leads to not printing the "You are now marked
as Available" line when reconnecting, leaving the "You are now offline"
as the last message in the chat.

The contact presence at this point is already correct, so we can
just bypass checking the connection and call it directly.

Reviewed-by: David Edmundson
BUG: 307827
FIXED-IN: 0.5.1