Removal of 'Typing...' in chat window title to avoid confusing UX
When a user is talking to multiple people, it is confusing to have 'Typing... Bob', when it's actually Fred in another tab who is typing.

The presence of the both the chat tab status-icon change combined with the tab name text colour change is sufficient to denote activity and removes this confusion which is exacerbated with the more chat tabs that you have open.