Use KTp::ContactList instead of KPeople
This means that we don't have direct dependency on KPeople and it also
solves fallback for us.

The code has been rewritten a lot, instead of waiting for the models to
populate and then merging them, we can now do this on-the-fly, as the
rows are being added to one or the other.