Add AdiumThemeViewProxy to allow the theme javascript to report when its finished initializing.
It does this by replacing the body onload event with a call to the Qt bridge
signal.

To make this compatible with various themes, during the template load
phase it finds the body tag, extracts the current onload event, saves any
javascript and replaces it with the call to the proxy object.

Then when javascript finishes initialising and tells Qt it's done,
and Qt can call the original event javascript.