Text-UI breaks "Per window" keyboard layout switching
This bug happens when the user has switched the keyboard layout of a chat tab and then switched to another application with another language, And he received a message or an "is typing" chat status, the language is then changed to the language of the Chat Tab that he was previously in, which is of course is an annoying behavior.

The reason of this behavior is that the restoreKeyboardLayout method is called inside the onCurrentIndexChanged without making sure that the Chat Window is active.

REVIEW: 117729
BUG: 333742