Revert "[lib] KDELibs4Support is not used at all, let's just remove it"
Somehow and I don't know how, it built at first, but that must have
been a mistake on my part as the lib/ is full of kdelibs4support usage

So I'm reverting this for now and will commit again once all the stuff
is ported away.

Sorry for the noise.