Fix kcm_ktp_chat_messages appearing in global search
The exec is not needed when we load it as a plugin and it is defunct anyways, because kcmshell4 does not work when using KF5 compiled KCMs.
Systemettings will skip KCMs in the deprecate dKServiceTypeTrader codepath that do not have an exec defined.


(cherry picked from commit f4df6eaf4a2eaaab66a7b91753fce1204ee9042c)